requirements.txt
# Core dependencies
openai>=1.0.0
networkx>=3.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Data and ML
datasets>=2.0.0

# Environment management
python-dotenv>=0.19.0
